agent sessions - consolidate opening actions (fix #280767)
