+++ src/vs/workbench/contrib/chat/browser/agentSessions/agentSessionsView.ts
@@ -4,17 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import './media/agentsessionsview.css';
-import { Codicon } from '../../../../../base/common/codicons.js';
-import { localize, localize2 } from '../../../../../nls.js';
-import { ContextKeyExpr, IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
-import { SyncDescriptor } from '../../../../../platform/instantiation/common/descriptors.js';
-import { Registry } from '../../../../../platform/registry/common/platform.js';
-import { registerIcon } from '../../../../../platform/theme/common/iconRegistry.js';
+import { localize } from '../../../../../nls.js';
+import { IContextKeyService } from '../../../../../platform/contextkey/common/contextkey.js';
 import { IViewPaneOptions, ViewPane } from '../../../../browser/parts/views/viewPane.js';
-import { ViewPaneContainer } from '../../../../browser/parts/views/viewPaneContainer.js';
-import { IViewContainersRegistry, Extensions as ViewExtensions, ViewContainerLocation, IViewsRegistry, IViewDescriptor, IViewDescriptorService } from '../../../../common/views.js';
+import { IViewDescriptorService } from '../../../../common/views.js';
 import { ChatContextKeys } from '../../common/chatContextKeys.js';
-import { ChatConfiguration } from '../../common/constants.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
 import { IContextMenuService } from '../../../../../platform/contextview/browser/contextView.js';
 import { IHoverService } from '../../../../../platform/hover/browser/hover.js';
@@ -36,7 +30,7 @@ import { DeferredPromise } from '../../../../../base/common/async.js';
 import { Event } from '../../../../../base/common/event.js';
 import { MutableDisposable } from '../../../../../base/common/lifecycle.js';
 import { getActionBarActions } from '../../../../../platform/actions/browser/menuEntryActionViewItem.js';
-import { AGENT_SESSIONS_VIEW_ID, AGENT_SESSIONS_VIEW_CONTAINER_ID, AgentSessionProviders } from './agentSessions.js';
+import { AgentSessionProviders } from './agentSessions.js';
 import { AgentSessionsFilter } from './agentSessionsFilter.js';
 import { AgentSessionsControl } from './agentSessionsControl.js';
 import { IAgentSessionsService } from './agentSessionsService.js';
@@ -88,7 +82,6 @@ export class AgentSessionsView extends ViewPane {
 				() => didResolve.p
 			);
 		}));
-
 	}
 
 	protected override renderBody(container: HTMLElement): void {
@@ -186,6 +179,7 @@ export class AgentSessionsView extends ViewPane {
 			}
 		}
 
+		// Install more
 		const installMenuActions = this.menuService.getMenuActions(MenuId.AgentSessionsInstallActions, this.scopedContextKeyService, { shouldForwardArgs: true });
 		const installActionBar = getActionBarActions(installMenuActions, () => true);
 		if (installActionBar.primary.length > 0) {
@@ -252,42 +246,3 @@ export class AgentSessionsView extends ViewPane {
 		this.sessionsControl?.focus();
 	}
 }
-
-//#region View Registration
-
-const chatAgentsIcon = registerIcon('chat-sessions-icon', Codicon.commentDiscussionSparkle, 'Icon for Agent Sessions View');
-
-const AGENT_SESSIONS_VIEW_TITLE = localize2('agentSessions.view.label', "Agents");
-
-const agentSessionsViewContainer = Registry.as<IViewContainersRegistry>(ViewExtensions.ViewContainersRegistry).registerViewContainer({
-	id: AGENT_SESSIONS_VIEW_CONTAINER_ID,
-	title: AGENT_SESSIONS_VIEW_TITLE,
-	icon: chatAgentsIcon,
-	ctorDescriptor: new SyncDescriptor(ViewPaneContainer, [AGENT_SESSIONS_VIEW_CONTAINER_ID, { mergeViewWithContainerWhenSingleView: true }]),
-	storageId: AGENT_SESSIONS_VIEW_CONTAINER_ID,
-	hideIfEmpty: true,
-	order: 6,
-}, ViewContainerLocation.AuxiliaryBar);
-
-const agentSessionsViewDescriptor: IViewDescriptor = {
-	id: AGENT_SESSIONS_VIEW_ID,
-	containerIcon: chatAgentsIcon,
-	containerTitle: AGENT_SESSIONS_VIEW_TITLE.value,
-	singleViewPaneContainerTitle: AGENT_SESSIONS_VIEW_TITLE.value,
-	name: AGENT_SESSIONS_VIEW_TITLE,
-	canToggleVisibility: false,
-	canMoveView: true,
-	openCommandActionDescriptor: {
-		id: AGENT_SESSIONS_VIEW_ID,
-		title: AGENT_SESSIONS_VIEW_TITLE
-	},
-	ctorDescriptor: new SyncDescriptor(AgentSessionsView),
-	when: ContextKeyExpr.and(
-		ChatContextKeys.Setup.hidden.negate(),
-		ChatContextKeys.Setup.disabled.negate(),
-		ContextKeyExpr.equals(`config.${ChatConfiguration.AgentSessionsViewLocation}`, 'single-view'),
-	)
-};
-Registry.as<IViewsRegistry>(ViewExtensions.ViewsRegistry).registerViews([agentSessionsViewDescriptor], agentSessionsViewContainer);
-
-//#endregion
